Replaced boost program_options with inih library Reworking project structure Added files to git:     dependency headers     resource file     3rd party .lib files     3rd party .dll files

+++ RetroGraphLib/Monitors.h
@@ -57,7 +57,7 @@ public:
     Monitors& operator=(Monitors&&) = delete;
 
     const std::vector<MonitorData>& getMonitorData() const { return m_monitors; }
-    size_t getNumMonitors() const { return m_monitors.size(); }
+    int getNumMonitors() const { return static_cast<int>(m_monitors.size()); }
 
     int32_t getWidth(size_t i) const { return m_monitors[i].width; }
     int32_t getHeight(size_t i) const { return m_monitors[i].height; }
